[FETCH ALBUMS] Add presenter implementation

+++ app/src/main/java/com/ccreusot/albums/entities/Photo.kt
@@ -1,3 +1,3 @@
 package com.ccreusot.albums.entities
 
-data class Photo(val id: Int, val title: String, val url: String, val thumbnailUrl: String)
+data class Photo(val id: Long, val title: String, val url: String, val thumbnailUrl: String)
